Add files via upload
Improved version of original version.
quote addition to All_QUotes.csv if it exist becomes optional
chooses random quote with owner from master file All_Quotes.csv

+++ randomQuoting.py
@@ -1,16 +1,26 @@
 import pandas as pd
+import numpy as np
 from quoteWriting import get_quote
 import os.path
 
 if os.path.isfile('All_Quotes.csv'):
-    print('File Exists')
-    dt = pd.read_csv('All_Quotes.csv', sep=';', index_col=0)
-    new_dt = get_quote()
-    dt = dt.append(new_dt)
-    dt.to_csv('All_Quotes.csv', sep=';')
+    inp = input('Do you want to add new quote (y/n): ')
+    if inp == 'y':
+        print('File Exists')
+        dt = pd.read_csv('All_Quotes.csv', sep=';', index_col=0)
+        new_dt = get_quote()
+        dt = dt.append(new_dt)
+        dt.to_csv('All_Quotes.csv', sep=';')
+    elif inp == 'n':
+        dt = pd.read_csv('All_Quotes.csv', sep=';', index_col=0)
+    else:
+        print('Wrong input')
+        dt = []
 else:
     print('File is created')
     dt = get_quote()
     dt.to_csv('All_Quotes.csv', sep=';')
 
-
+idx = np.random.randint(0, len(dt))
+r = dt.iloc[idx]
+print(r)
